requirements.txt
pytest==6.2.2
requests==2.25.1
faker==5.0.0
selenium==3.141.0
mitmproxy==6.0.2
allure-pytest==2.9.42